+++ NEWS.md
@@ -5,6 +5,8 @@ NEWS
 
 # TOSTER v0.9.0
 - Added `perm_t_test` function to allow for permutation tests for equivalence using TOST
+- Update `brunner_munzel` function to allow TOST directly
+- Update functions to disallow `paired = TRUE` when formula method utilized.
 
 # TOSTER v0.8.7
 - Update documentation to make it clear what the "eqb" argument does within the `wilcox_TOST` function.